+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Change Log
-## 0.1.0 - 2020-11-20
+## [0.2.0] - 2020-12-11
+
+### Added
+
+- Undoing moves.
+- Option to quit game before game finished.
+- Option to set board size.
+- Option to set number of players to either 2 or 3.
+
+### Changed
+
+- Starting piece distribution, both for code elegance and to make 3 player games fun.
+- The state now stores the history (to support undoing moves).
+
+### Fixed
+
+- Out-of-bounds checking function.
+
+## [0.1.0] - 2020-12-10
+
 ### Added
 - Implementation of Othello rules.
 - User interaction via terminal.